requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.29.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
cryptography==41.0.7
httpx==0.25.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
mangum==0.19.0
firebase-admin==6.2.0

# Music extraction and analysis (optional)
yt-dlp==2023.12.30
librosa==0.10.1
numpy==1.24.3
requests==2.31.0